lib/core: Move temperature API from hwtest
This commit moves APIs related to temperature from test/hwtest to
lib.
