+++ lib/core/CMakeLists.txt
@@ -5,6 +5,8 @@ if (CONFIG_BOARD_SCSAT1_MAIN OR CONFIG_BOARD_SCSAT1_ADCS)
 	zephyr_library()
 	zephyr_library_sources(wdog.c)
 	zephyr_library_sources_ifdef(CONFIG_SC_LIB_CORE_MAIN pwrctrl_main.c)
+	zephyr_library_sources_ifdef(CONFIG_SC_LIB_CORE_MAIN temp_main.c)
 	zephyr_library_sources_ifdef(CONFIG_SC_LIB_CORE_ADCS pwrctrl_adcs.c)
+	zephyr_library_sources_ifdef(CONFIG_SC_LIB_CORE_ADCS temp_adcs.c)
 	zephyr_include_directories(.)
 endif()
