+++ lib/core/temp_adcs.h
