+++ lib/core/temp_main.h
