+++ tests/hwtest/main/CMakeLists.txt
@@ -11,7 +11,6 @@ project(scsat1-main)
 target_sources(app PRIVATE src/main.c)
 target_sources(app PRIVATE src/mgnm.c )
 target_sources(app PRIVATE src/mgnm_test.c )
-target_sources(app PRIVATE src/temp.c)
 target_sources(app PRIVATE src/temp_test.c)
 target_sources(app PRIVATE src/cv.c)
 target_sources(app PRIVATE src/cv_test.c)
